Integrate mongodb sample into asp.net sample
