--- samples/Samples.MongoDb/Properties/launchSettings.json
@@ -1,26 +1,0 @@
-{
-  "profiles": {
-    "Samples.MongoDb": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{918728DD-259F-4A6A-AC2B-B85E1B658318}",
-        "CORECLR_PROFILER_PATH": "$(ProjectDir)$(OutputPath)profiler-lib\\OpenTelemetry.AutoInstrumentation.ClrProfiler.Native.dll",
-
-        "OTEL_DOTNET_TRACER_HOME": "$(ProjectDir)$(OutputPath)profiler-lib",
-        "OTEL_INTEGRATIONS": "$(ProjectDir)$(OutputPath)profiler-lib\\integrations.json",
-        "OTEL_PROFILER_PROCESSES": "Samples.MongoDb.exe",
-
-        "OTEL_EXPORTER": "jaeger",
-        "OTEL_EXPORTER_JAEGER_AGENT_HOST": "localhost",
-        "OTEL_EXPORTER_JAEGER_AGENT_PORT": "6831",
-
-        "OTEL_DUMP_ILREWRITE_ENABLED": "0",
-        "OTEL_TRACE_CALLTARGET_ENABLED": "1",
-        "OTEL_CLR_ENABLE_INLINING": "1"
-      }
-    }
-  }
-}
